Adding support to request GTFS RT updates from site with username and password access requirements
Added the ability for GtfsRealtimeHttpTripUpdateSource to check the
routers-config.json for username and password. This can then be passed
through the system to HttpUtils which will add the appropriate basic
authenticator and be able to request updates from the site with the
provided credentials.

An additional change to the pom was required to support the
authentication, adding a dependency to commons-codec.